// src/generated/api/returns/models/inspectReturnItemDtoCondition.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type InspectReturnItemDtoCondition =
  (typeof InspectReturnItemDtoCondition)[keyof typeof InspectReturnItemDtoCondition];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InspectReturnItemDtoCondition = {
  SELLABLE: 'SELLABLE',
  DAMAGED: 'DAMAGED',
  MISSING: 'MISSING',
} as const;