// src/generated/api/returns/models/returnItemDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */
import type { ReturnItemDtoItemType } from './returnItemDtoItemType';
import type { ReturnItemDtoCondition } from './returnItemDtoCondition';
import type { ReturnItemDtoDisposition } from './returnItemDtoDisposition';

export interface ReturnItemDto {
  /** @pattern ^[1-9][0-9]*$ */
  id: string;
  /** @pattern ^[1-9][0-9]*$ */
  orderItemId: string;
  sku: string;
  productName: string;
  variantName: string;
  itemType: ReturnItemDtoItemType;
  quantity: number;
  unitPrice: string;
  /** @nullable */
  condition?: ReturnItemDtoCondition;
  /** @nullable */
  disposition?: ReturnItemDtoDisposition;
  restockQty: number;
  /** @nullable */
  refundCap?: string | null;
  /** @nullable */
  note?: string | null;
}